Extract commands into modules
