+++ commands/index.js
@@ -1,0 +1,8 @@
+const activateFn = require('./activate');
+const createEnv = require('./create-environment');
+const deploy = require('./deploy');
+const init = require('./init');
+const newFn = require('./new-function');
+const start = require('./start');
+
+module.exports = { activateFn, createEnv, deploy, init, newFn, start };
